+++ kitchenpos-order/src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -8,7 +8,7 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
-import kitchenpos.common.domain.Quantity;
+import kitchenpos.domain.Quantity;
 
 @Entity
 public class OrderLineItem {
