+++ kitchenpos-order/src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -3,12 +3,12 @@ package kitchenpos.order.domain;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.Collections;
-import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableFactory;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import kitchenpos.table.domain.OrderTable;
 
 class OrderTest {
